init: 프로젝트 초기 생성 및 라이브러리 설치
