+++ next.config.mjs
@@ -1,6 +1,9 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   reactStrictMode: true,
-}
+  compiler: {
+    styledComponents: true,
+  },
+};
 
-module.exports = nextConfig
+export default nextConfig;
