+++ src/pages/_document.tsx
@@ -1,0 +1,27 @@
+import { Html, Head, Main, NextScript } from 'next/document';
+import Link from 'next/link';
+
+export default function Document() {
+  return (
+    <Html lang="ko">
+      <Head>
+        <Link
+          rel="stylesheet"
+          type="text/css"
+          href="https://cdn.jsdelivr.net/gh/orioncactus/pretendard/dist/web/static/pretendard.css"
+        />
+        <script
+          defer
+          src="https://t1.kakaocdn.net/kakao_js_sdk/2.6.0/kakao.min.js"
+          integrity="sha384-6MFdIr0zOira1CHQkedUqJVql0YtcZA1P0nbPrQYJXVJZUkTk/oX4U9GhUIs3/z8"
+          crossOrigin="anonymous"
+        ></script>
+      </Head>
+      <body>
+        <Main />
+        <NextScript />
+        <div id="modal-root"></div>
+      </body>
+    </Html>
+  );
+}
